requirements.txt
chainlit
openai
python-dotenv
langchain_openai 
langchain_core
python-dotenv
langchain_community
langserve
-e .